Move Novation USB Driver logic to Helpers.DriverChecker
