+++ Apollo/Core/App.cs
@@ -2,7 +2,6 @@
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
-using System.Linq;
 using System.Runtime.InteropServices;
 
 using Avalonia;
@@ -30,54 +29,6 @@ namespace Apollo.Core {
         public static readonly KeyModifiers ControlKey = RuntimeInformation.IsOSPlatform(OSPlatform.OSX)? KeyModifiers.Meta : KeyModifiers.Control;
 
         public static bool Dragging = false;
-
-        class DriverVersion: IComparable {
-            public int Minor { get; private set; }
-            public int Build { get; private set; }
-
-            public DriverVersion(int[] version) {
-                Minor = version[0];
-                Build = version[1];
-            }
-
-            public int CompareTo(object obj) {
-                DriverVersion that = (DriverVersion)obj;
-
-                if (this.Minor == that.Minor) {
-                    if (this.Build == that.Build) return 0;
-                    return (this.Build < that.Build)? -1 : 1;
-                }
-                
-                return (this.Minor < that.Minor)? -1 : 1;
-            }
-
-            public static bool operator <(DriverVersion a, DriverVersion b)
-                => a.CompareTo(b) == -1;
-
-            public static bool operator >(DriverVersion a, DriverVersion b) 
-                => a.CompareTo(b) == 1;
-        }
-
-        MessageWindow CreateDriverError(bool IsOldVersion) {
-            MessageWindow ret = new MessageWindow(
-                (IsOldVersion
-                    ? "Apollo Studio requires a newer version of the Novation USB Driver than is\n" +
-                      "installed on your computer.\n\n"
-                    : "Apollo Studio requires the Novation USB Driver which isn't installed on your\n" +
-                      "computer.\n\n"
-                ) +
-                "Please install at least version 2.15.5 of the driver before launching Apollo\n" +
-                "Studio.",
-                new string[] {"Download Driver", "OK"}
-            );
-
-            ret.Completed.Task.ContinueWith(result => {
-                if (result.Result == "Download Driver")
-                    URL("https://customer.novationmusic.com/sites/customer/files/downloads/Novation%20USB%20Driver-2.15.5.exe");
-            });
-
-            return ret;
-        }
 
         public static bool WindowKey(Window sender, KeyEventArgs e) {
             if (e.KeyModifiers == ControlKey) {
@@ -145,29 +96,11 @@ namespace Apollo.Core {
 
             if (Args.Length > 0 && Args[0] == "--update") lifetime.MainWindow = new UpdateWindow();
             else {
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows)) { // USB Driver check
-                    IEnumerable<DriverVersion> a = (from j in Directory.GetDirectories(Environment.ExpandEnvironmentVariables(@"%SystemRoot%\System32\DriverStore\FileRepository\"))
-                        where Path.GetFileName(j).StartsWith("nvnusbaudio.inf")
-                        select new DriverVersion((
-                            from i in File.ReadAllLines(Path.Combine(j, "nvnusbaudio.inf"))
-                                where i.StartsWith("DriverVer=")
-                                select i
-                        ).First().Substring(10).Split(',')[1].Split('.').Where((x, i) => i == 1 || i == 3).Select(x => Convert.ToInt32(x)).ToArray())
-                    );
+                if (!DriverChecker.Run(out MessageWindow driverError)) {
+                    lifetime.MainWindow = driverError;
 
-                    if (a.Count() == 0) {
-                        lifetime.MainWindow = CreateDriverError(false);
-
-                        base.OnFrameworkInitializationCompleted();
-                        return;
-                    }
-
-                    if (a.Max() < new DriverVersion(new int[] {15, 5})) {
-                        lifetime.MainWindow = CreateDriverError(true);
-
-                        base.OnFrameworkInitializationCompleted();
-                        return;
-                    }
+                    base.OnFrameworkInitializationCompleted();
+                    return;
                 }
 
                 if (!AbletonConnector.Connected) {
